requirements.txt
requests==2.27.1
beautifulsoup4==4.10.0
bs4==0.0.1
pushbullet.py==0.12.0
lxml==4.8.0
python-dotenv==0.19.2
